+++ automation/dictionary/index.js
@@ -1,4 +1,7 @@
-const randomPhoneNum = () => Number(new Date().getTime()).toString().slice(0, 10);
+const randomPhoneNum = () =>
+  Number(new Date().getTime())
+    .toString()
+    .slice(0, 10);
 const randomId = randomPhoneNum();
 // TODO
 // Reason Lists - Needs extra logic for creation
@@ -25,15 +28,15 @@ const randomId = randomPhoneNum();
 // Agent State Monitoring - ready to be worked on
 
 const pages = {
-////////////////////////User Management///////////////////////////
-  'User': {
+  ////////////////////////User Management///////////////////////////
+  User: {
     specs: {
       create: {
         parametersToInsert: [
           {
             emailInput: `remail-${randomId}@gmail.com`,
-            roleList: "Platform User",
-            tenantList: "Agent"
+            roleList: 'Platform User',
+            tenantList: 'Agent'
           }
         ]
       },
@@ -42,7 +45,7 @@ const pages = {
           {
             firstNameInput: `updatedRName-${randomId}`,
             lastNameInput: `updatedRDesc-${randomId}`,
-            roleList: "Supervisor",
+            roleList: 'Supervisor'
           }
         ]
       },
@@ -51,110 +54,110 @@ const pages = {
       }
     },
     navigation: {
-      mainBar: "userManagementMenu",
-      subMainBar: "navigationLinkUsers"
+      mainBar: 'userManagementMenu',
+      subMainBar: 'navigationLinkUsers'
     },
-    whichCatagoryToSearch: "searchEmailColumn",
+    whichCatagoryToSearch: 'searchEmailColumn',
     updateSearchValue: `remail-${randomId}@gmail.com`
   },
-  'Group': {
-        specs: {
-          create: {
-            parametersToInsert: [
-              {
-                nameInput: `rName-${randomId}`,
-                descriptionInput: `rDesc-${randomId}`
-              }
-            ]
-          },
-          update: {
-            parametersToInsert: [
-              {
-                nameInput: `updatedRName-${randomId}`,
-                descriptionInput: `updatedRDesc-${randomId}`
-              }
-            ]
-          },
-          read: {
-            parameters: {}
+  Group: {
+    specs: {
+      create: {
+        parametersToInsert: [
+          {
+            nameInput: `rName-${randomId}`,
+            descriptionInput: `rDesc-${randomId}`
           }
-        },
-        navigation: {
-          mainBar: "userManagementMenu",
-          subMainBar: "navigationLinkGroups"
-        },
-        whichCatagoryToSearch: "searchNameColumn",
-        updateSearchValue: `rName-${randomId}`
+        ]
       },
-  'Skill': {
-        specs: {
-          create: {
-            parametersToInsert: [
-              {
-                nameInput: `rName-${randomId}`,
-                descriptionInput: `rDesc-${randomId}`
-                // hasProficiencyToggle: ""
-              }
-            ]
-          },
-          update: {
-            parametersToInsert: [
-              {
-                nameInput: `updatedRName-${randomId}`,
-                descriptionInput: `updatedRDesc-${randomId}`,
-              }
-            ]
-          },
-          read: {
-            parameters: {}
+      update: {
+        parametersToInsert: [
+          {
+            nameInput: `updatedRName-${randomId}`,
+            descriptionInput: `updatedRDesc-${randomId}`
           }
-        },
-        navigation: {
-          mainBar: "userManagementMenu",
-          subMainBar: "navigationLinkSkills"
-        },
-        whichCatagoryToSearch: "searchNameColumn",
-        updateSearchValue: `rName-${randomId}`
+        ]
       },
-  'Role': {
-        specs: {
-          create: {
-            parametersToInsert: [
-              {
-                nameInput: `rName-${randomId}`,
-                descriptionInput: `rDesc-${randomId}`,
-                // sharedToggle: ""
-                }
-            ]
-          },
-          update: {
-            parametersToInsert: [
-              {
-                nameInput: `updatedRName-${randomId}`,
-                descriptionInput: `updatedRDesc-${randomId}`,
-              }
-            ]
-          },
-          read: {
-            parameters: {}
+      read: {
+        parameters: {}
+      }
+    },
+    navigation: {
+      mainBar: 'userManagementMenu',
+      subMainBar: 'navigationLinkGroups'
+    },
+    whichCatagoryToSearch: 'searchNameColumn',
+    updateSearchValue: `rName-${randomId}`
+  },
+  Skill: {
+    specs: {
+      create: {
+        parametersToInsert: [
+          {
+            nameInput: `rName-${randomId}`,
+            descriptionInput: `rDesc-${randomId}`
+            // hasProficiencyToggle: ""
           }
-        },
-        navigation: {
-          mainBar: "userManagementMenu",
-          subMainBar: "navigationLinkRoles"
-        },
-        whichCatagoryToSearch: "searchNameColumn",
-        updateSearchValue: `rName-${randomId}`
+        ]
       },
-  'Reason': {
+      update: {
+        parametersToInsert: [
+          {
+            nameInput: `updatedRName-${randomId}`,
+            descriptionInput: `updatedRDesc-${randomId}`
+          }
+        ]
+      },
+      read: {
+        parameters: {}
+      }
+    },
+    navigation: {
+      mainBar: 'userManagementMenu',
+      subMainBar: 'navigationLinkSkills'
+    },
+    whichCatagoryToSearch: 'searchNameColumn',
+    updateSearchValue: `rName-${randomId}`
+  },
+  Role: {
+    specs: {
+      create: {
+        parametersToInsert: [
+          {
+            nameInput: `rName-${randomId}`,
+            descriptionInput: `rDesc-${randomId}`
+            // sharedToggle: ""
+          }
+        ]
+      },
+      update: {
+        parametersToInsert: [
+          {
+            nameInput: `updatedRName-${randomId}`,
+            descriptionInput: `updatedRDesc-${randomId}`
+          }
+        ]
+      },
+      read: {
+        parameters: {}
+      }
+    },
+    navigation: {
+      mainBar: 'userManagementMenu',
+      subMainBar: 'navigationLinkRoles'
+    },
+    whichCatagoryToSearch: 'searchNameColumn',
+    updateSearchValue: `rName-${randomId}`
+  },
+  Reason: {
     specs: {
       create: {
         parametersToInsert: [
           {
             nameInput: `rName-${randomId}`,
             descriptionInput: `rDesc-${randomId}`,
-            externalIdInput: "",
-            sharedToggle: ""
+            externalIdInput: '',
+            sharedToggle: ''
           }
         ]
       },
@@ -163,7 +166,7 @@ const pages = {
           {
             nameInput: `updatedRName-${randomId}`,
             descriptionInput: `updatedRDesc-${randomId}`,
-            externalIdInput: "",
+            externalIdInput: ''
           }
         ]
       },
@@ -172,63 +175,63 @@ const pages = {
       }
     },
     navigation: {
-      mainBar: "userManagementMenu",
-      subMainBar: "navigationLinkReasons"
+      mainBar: 'userManagementMenu',
+      subMainBar: 'navigationLinkReasons'
     },
-    whichCatagoryToSearch: "searchNameColumn",
+    whichCatagoryToSearch: 'searchNameColumn',
     updateSearchValue: `rName-${randomId}`
   },
-// 'Reasonlist': {
-//   specs: {
-//     create: {
-//       parametersToInsert: [
-//         {
-//           nameInput: `rName-${randomId}`,
-//           descriptionInput: `rDesc-${randomId}`,
-//           externalIdInput: "",
-//           //sharedToggle: "",
-//           //isDefaultToggle: ""
-//           }
-//       ]
-//     },
-//     update: {
-//       parametersToInsert: [
-//         {
-//           nameInput: `updatedRName-${randomId}`,
-//           descriptionInput: `updatedRDesc-${randomId}`,
-//           externalIdInput: "",
-//           //sharedToggle: "",
-//           //isDefaultToggle: ""
-//         }
-//       ]
-//     },
-//     read: {
-//       parameters: {}
-//     }
-//   },
-//   navigation: {
-//     mainBar: "userManagementMenu",
-//     subMainBar: "navigationLinkReasonsList"
-//   },
-//   whichCatagoryToSearch: "searchNameColumn",
-//   updateSearchValue: `rName-${randomId}`
-// },
-//////////////////////////Configuration///////////////////////////
-  'List': {
+  // 'Reasonlist': {
+  //   specs: {
+  //     create: {
+  //       parametersToInsert: [
+  //         {
+  //           nameInput: `rName-${randomId}`,
+  //           descriptionInput: `rDesc-${randomId}`,
+  //           externalIdInput: "",
+  //           //sharedToggle: "",
+  //           //isDefaultToggle: ""
+  //           }
+  //       ]
+  //     },
+  //     update: {
+  //       parametersToInsert: [
+  //         {
+  //           nameInput: `updatedRName-${randomId}`,
+  //           descriptionInput: `updatedRDesc-${randomId}`,
+  //           externalIdInput: "",
+  //           //sharedToggle: "",
+  //           //isDefaultToggle: ""
+  //         }
+  //       ]
+  //     },
+  //     read: {
+  //       parameters: {}
+  //     }
+  //   },
+  //   navigation: {
+  //     mainBar: "userManagementMenu",
+  //     subMainBar: "navigationLinkReasonsList"
+  //   },
+  //   whichCatagoryToSearch: "searchNameColumn",
+  //   updateSearchValue: `rName-${randomId}`
+  // },
+  //////////////////////////Configuration///////////////////////////
+  List: {
     specs: {
       create: {
         parametersToInsert: [
           {
             nameInput: `rName-${randomId}`,
             //sharedToggle: "",
-            listTypeList: "Disposition Codes Type"
+            listTypeList: 'Disposition Codes Type'
           }
         ]
       },
       update: {
         parametersToInsert: [
           {
-            nameInput: `updatedRName-${randomId}`,
+            nameInput: `updatedRName-${randomId}`
             //sharedToggle: ""
           }
         ]
@@ -238,31 +241,31 @@ const pages = {
       }
     },
     navigation: {
-      mainBar: "configurationMenu",
-      subMainBar: "navigationLinkLists"
+      mainBar: 'configurationMenu',
+      subMainBar: 'navigationLinkLists'
     },
-    whichCatagoryToSearch: "searchNameColumn",
+    whichCatagoryToSearch: 'searchNameColumn',
     updateSearchValue: `rName-${randomId}`
   },
-  'Sla': {
+  Sla: {
     specs: {
       create: {
         parametersToInsert: [
           {
             nameInput: `rName-${randomId}`,
             descriptionInput: `rDesc-${randomId}`,
-            sharedToggle: ""
+            sharedToggle: ''
           }
         ]
       },
-      createVersion:{
+      createVersion: {
         parametersToInsert: [
           {
             //Name and Description modal elements defined in webElements.js
             //Name and Description data added via inputFormDataForModal in common.js
-            thresholdInput: "20",
-            abandonTypeRadio: "ignore-abandons",
-            abandonThresholdInput: "30"
+            thresholdInput: '20',
+            abandonTypeRadio: 'ignore-abandons',
+            abandonThresholdInput: '30'
           }
         ]
       },
@@ -279,10 +282,10 @@ const pages = {
       }
     },
     navigation: {
-      mainBar: "configurationMenu",
-      subMainBar: "navigationLinkSlas"
+      mainBar: 'configurationMenu',
+      subMainBar: 'navigationLinkSlas'
     },
-    whichCatagoryToSearch: "searchNameColumn",
+    whichCatagoryToSearch: 'searchNameColumn',
     updateSearchValue: `rName-${randomId}`
   },
   // 'Email Template': {
@@ -317,7 +320,7 @@ const pages = {
             descriptionInput: `rDesc-${randomId}`,
             channelList: 1,
             flowList: 1,
-            valueInput: "+15063001234"
+            valueInput: '+15063001234'
           }
         ]
       },
@@ -328,7 +331,7 @@ const pages = {
             descriptionInput: `updatedRDesc-${randomId}`,
             channelList: 0,
             flowList: 2,
-            valueInput: "+15063004321"
+            valueInput: '+15063004321'
           }
         ]
       },
@@ -337,10 +340,10 @@ const pages = {
       }
     },
     navigation: {
-      mainBar: "configurationMenu",
-      subMainBar: "navigationLinkOutboundIdentifiers"
+      mainBar: 'configurationMenu',
+      subMainBar: 'navigationLinkOutboundIdentifiers'
     },
-    whichCatagoryToSearch: "searchNameColumn",
+    whichCatagoryToSearch: 'searchNameColumn',
     updateSearchValue: `rName-${randomId}`
   },
   'Outbound Identifier List': {
@@ -349,7 +352,7 @@ const pages = {
         parametersToInsert: [
           {
             nameInput: `rName-${randomId}`,
-            descriptionInput: `rDesc-${randomId}`,
+            descriptionInput: `rDesc-${randomId}`
           }
         ]
       },
@@ -357,7 +360,7 @@ const pages = {
         parametersToInsert: [
           {
             nameInput: `updatedRName-${randomId}`,
-            descriptionInput: `updatedRDesc-${randomId}`,
+            descriptionInput: `updatedRDesc-${randomId}`
           }
         ]
       },
@@ -366,10 +369,10 @@ const pages = {
       }
     },
     navigation: {
-      mainBar: "configurationMenu",
-      subMainBar: "navigationLinkOutboundIdentifierLists"
+      mainBar: 'configurationMenu',
+      subMainBar: 'navigationLinkOutboundIdentifierLists'
     },
-    whichCatagoryToSearch: "searchNameColumn",
+    whichCatagoryToSearch: 'searchNameColumn',
     updateSearchValue: `rName-${randomId}`
   },
   // 'Chat Widget': {
@@ -435,7 +438,7 @@ const pages = {
           {
             nameInput: `keyName-${randomId}`,
             descriptionInput: `keyDesc-${randomId}`,
-            roleList: "Administrator"
+            roleList: 'Administrator'
           }
         ]
       },
@@ -444,7 +447,7 @@ const pages = {
           {
             nameInput: `updatedKeyName-${randomId}`,
             descriptionInput: `updatedKeyDesc-${randomId}`,
-            roleList: "Agent"
+            roleList: 'Agent'
           }
         ]
       },
@@ -454,12 +457,12 @@ const pages = {
       }
     },
     navigation: {
-      mainBar: "configurationMenu",
-      subMainBar: "navigationLinkApiKeys"
+      mainBar: 'configurationMenu',
+      subMainBar: 'navigationLinkApiKeys'
     },
-    whichCatagoryToSearch: "searchNameColumn",
+    whichCatagoryToSearch: 'searchNameColumn',
     updateSearchValue: `keyName-${randomId}`,
-    deleteSearchValue: `updatedKeyName-${randomId}`,
+    deleteSearchValue: `updatedKeyName-${randomId}`
   },
   // 'Message Template': {
   //   specs: {
@@ -515,13 +518,13 @@ const pages = {
       }
     },
     navigation: {
-      mainBar: "configurationMenu",
-      subMainBar: "navigationLinkBusinessHours"
+      mainBar: 'configurationMenu',
+      subMainBar: 'navigationLinkBusinessHours'
     },
-    whichCatagoryToSearch: "searchNameColumn",
+    whichCatagoryToSearch: 'searchNameColumn',
     updateSearchValue: `rName-${randomId}`
   },
-/////////////////////////////////Flows//////////////////////////////
+  /////////////////////////////////Flows//////////////////////////////
   // 'Flow': {
   //   specs: {
   //     create: {
@@ -560,8 +563,8 @@ const pages = {
           {
             nameInput: `dMapName-${randomId}`,
             descriptionInput: `dMapDesc-${randomId}`,
-            interactionList: "Voice",
-            mappingList: "Contact Point",
+            interactionList: 'Voice',
+            mappingList: 'Contact Point',
             phoneValueInput: `+1${randomPhoneNum()}`,
             flowList: 1
           }
@@ -572,8 +575,8 @@ const pages = {
           {
             nameInput: `updatedDMapName-${randomId}`,
             descriptionInput: `updatedDMapDesc-${randomId}`,
-            interactionList: "Sms",
-            mappingList: "Customer",
+            interactionList: 'Sms',
+            mappingList: 'Customer',
             phoneValueInput: `+1${randomPhoneNum()}`,
             flowList: 2,
             flowVersionList: 1
@@ -585,13 +588,13 @@ const pages = {
       }
     },
     navigation: {
-      mainBar: "flowsMenu",
-      subMainBar: "navigationLinkDispatchMappings"
+      mainBar: 'flowsMenu',
+      subMainBar: 'navigationLinkDispatchMappings'
     },
-    whichCatagoryToSearch: "searchNameColumn",
+    whichCatagoryToSearch: 'searchNameColumn',
     updateSearchValue: `dMapName-${randomId}`
   },
-/////////////////////////////Reporting/////////////////////////////
+  /////////////////////////////Reporting/////////////////////////////
   'Data Access Report': {
     specs: {
       create: {
@@ -599,9 +602,9 @@ const pages = {
           {
             nameInput: `rName-${randomId}`,
             descriptionInput: `rDesc-${randomId}`,
-            reportTypeRadio: "realtime",
-            realtimeReportTypeRadio: "standard",
-            realtimeReportAutoComplete: "Agent State Table"
+            reportTypeRadio: 'realtime',
+            realtimeReportTypeRadio: 'standard',
+            realtimeReportAutoComplete: 'Agent State Table'
           }
         ]
       },
@@ -609,7 +612,7 @@ const pages = {
         parametersToInsert: [
           {
             nameInput: `updatedRName-${randomId}`,
-            descriptionInput: `updatedRDesc-${randomId}`,
+            descriptionInput: `updatedRDesc-${randomId}`
             //realtimeReportAutoComplete: "Interactions Completed Table"
           }
         ]
@@ -619,10 +622,10 @@ const pages = {
       }
     },
     navigation: {
-      mainBar: "reportingMenu",
-      subMainBar: "navigationLinkDataAccessReports"
+      mainBar: 'reportingMenu',
+      subMainBar: 'navigationLinkDataAccessReports'
     },
-    whichCatagoryToSearch: "searchNameColumn",
+    whichCatagoryToSearch: 'searchNameColumn',
     updateSearchValue: `rName-${randomId}`
   }
 };