+++ src/redux/modules/entities/lists/selectors.js
@@ -1,15 +1,12 @@
 /*
- * Copyright © 2015-2018 Serenova, LLC. All rights reserved.
+ * Copyright © 2015-2019 Serenova, LLC. All rights reserved.
  */
 
-import { Map } from 'immutable';
+import { createSelector } from 'reselect';
 import { getSelectedEntity } from '../selectors';
 
-export const getListTypeName = state =>
-  getSelectedEntity(state).getIn(['listType', 'name']);
+export const getListTypeName = state => getSelectedEntity(state).getIn(['listType', 'name']);
 
-export const getInitialUpdateFormValues = state =>
-  new Map({
-    name: getSelectedEntity(state).get('name'),
-    shared: getSelectedEntity(state).get('shared')
-  });
+export const getInitialUpdateFormValues = createSelector([getSelectedEntity], selectedEntity =>
+  selectedEntity.delete('active').delete('items')
+);